Add towns on unsearched roads from mileage targets
Since the positions of these twons are available in mileage targets, it's quite likely that they actually exist in the game. Until these stretches of roads are looked at manually, so that the towns can be either confirmed or removed, the mileage targets data is as good as it gets.