Swap the order of remark and ref in metadata spec
`ref` would probably mostly be useful for automated use. For example, if it contained a GeoNames URL, you could query that to try and determine the appropriate exonym for your localization.

By contrast, `remark` is mostly for human use. It might contain information that's important during editing, and it should preferably not require horizontal scrolling to read.

Right now this is a non-issue because `ref` is unused and always empty. But if and when that changes, it's better for `remark` to come before `ref`.