Add additional names found during 2025 Wyoming research
These are names that should perhaps not be shown by default (or at least not in the same way as regular scenery towns).

These labels are primarily for "parking" and "historic" features. I imagine it would be nice to show these in search results and on the map, but probably only on a higher zoom level than labels for scenery towns. For parking at nature locations, I considered the importance of such locations when deciding whether to categorize it as "parking" or as "nature" (e.g. Beaver Rim vs. Hell's Half Acre).

A few labels are for features of another kind. Newly added categories are "airport" and "industrial" because no existing category seemed to fit.