Produce all-towns GeoJSON from extra-labels
The maps project can't read the new metadata yet. The required change is simple and has indeed been ready for review for a couple of weeks now (https://github.com/nautofon/maps-nodejs/pull/3). But it's currently not known when it will actually be merged and made available on truckermudgeon.github.io.

Therefore, we still need to create the legacy all-towns dataset in the old format for the time being.

A simple new script achieves that by creating the legacy name/state properties from text/country. It also removes unneeded GeoJSON features and properties from the output and tries to achieve the same sort order as in the old file. Even so, many lines differ slightly from the old file to the new one. This is due to the changed policy on abbreviations and coordinate precision in metadata.